fix bug

+++ database/migrations/2025_07_12_192337_create_transaction_details_table.php
@@ -16,11 +16,11 @@ return new class extends Migration
             $table->foreignId('transaction_id')->constrained('transactions')->onDelete('cascade');
             $table->foreignId('part_id')->constrained('parts')->onDelete('cascade');
             $table->integer('qty');
-            $table->integer('subtotal');
+            $table->bigInteger('subtotal');
             $table->timestamps();
         });
     }
- 
+
     /**
      * Reverse the migrations.
      */
